Remove clipping in [0, 1]
Since the noise is being added after the preprocessing step, clipping the pixel values don't make any sense
In foolbox libraries, preprocessing is done after adding the perturbation and is the standard way!
Adding perturbation after preprocessing is not wrong, but it does not give you the correct perturbation norm value.